fix: #70 ProfileInfo 컴포넌트 server 컴포넌트로 수정, server session 값 확인해서 라우팅하도록 수정, image 오류 수정
